+++ src/ToolchainManager.ts
@@ -100,8 +100,9 @@ export interface IToolchian {
 
     /**
      * get gcc c/c++ compiler path for cpptools
+     * @param type c or c++ compiler
      */
-    getGccFamilyCompilerPathForCpptools(): string | undefined;
+    getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined;
 
     /**
      * get compiler internal defines (for static check)
@@ -472,7 +473,7 @@ class KeilC51 implements IToolchian {
         return new KeilC51();
     }
 
-    getGccFamilyCompilerPathForCpptools(): string | undefined {
+    getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined {
         //const gcc = File.fromArray([this.getToolchainDir().path, 'BIN', `C51${platform.exeSuffix()}`]);
         //return gcc.path;
         return undefined;
@@ -662,7 +663,7 @@ class SDCC implements IToolchian {
         }
     }
 
-    getGccFamilyCompilerPathForCpptools(): string | undefined {
+    getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined {
         //const gcc = File.fromArray([this.getToolchainDir().path, 'bin', `sdcc${platform.exeSuffix()}`]);
         //return gcc.path;
         return undefined;
@@ -909,7 +910,7 @@ class SDCC implements IToolchian {
 //         return new SDCC();
 //     }
 
-//     getGccFamilyCompilerPathForCpptools(): string | undefined {
+//     getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined {
 //         //const gcc = File.fromArray([this.getToolchainDir().path, 'bin', `sdcc${platform.exeSuffix()}`]);
 //         //return gcc.path;
 //         return undefined;
@@ -1117,7 +1118,7 @@ class COSMIC_STM8 implements IToolchian {
         return new COSMIC_STM8();
     }
 
-    getGccFamilyCompilerPathForCpptools(): string | undefined {
+    getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined {
         // not support
         return undefined;
     }
@@ -1449,7 +1450,7 @@ class AC5 implements IToolchian {
         return new AC5();
     }
 
-    getGccFamilyCompilerPathForCpptools(): string | undefined {
+    getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined {
         //const armccFile = File.fromArray([this.getToolchainDir().path, 'bin', `armcc${platform.exeSuffix()}`]);
         //return armccFile.path;
         return undefined;
@@ -1681,7 +1682,7 @@ class AC6 implements IToolchian {
         return new AC6();
     }
 
-    getGccFamilyCompilerPathForCpptools(): string | undefined {
+    getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined {
         const armclang = File.fromArray([this.getToolchainDir().path, 'bin', `armclang${platform.exeSuffix()}`]);
         return armclang.path;
     }
@@ -1911,8 +1912,10 @@ class GCC implements IToolchian {
         return new GCC();
     }
 
-    getGccFamilyCompilerPathForCpptools(): string | undefined {
-        const gcc = File.fromArray([this.getToolchainDir().path, 'bin', this.getToolPrefix() + `gcc${platform.exeSuffix()}`]);
+    getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined {
+        const gcc = File.from(
+            this.getToolchainDir().path, 'bin', 
+            this.getToolPrefix() + `${type == 'c++' ? 'g++' : 'gcc'}${platform.exeSuffix()}`);
         return gcc.path;
     }
 
@@ -2134,7 +2137,7 @@ class IARARM implements IToolchian {
         return new IARARM();
     }
 
-    getGccFamilyCompilerPathForCpptools(): string | undefined {
+    getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined {
         return undefined;
     }
 
@@ -2272,7 +2275,7 @@ class IARSTM8 implements IToolchian {
         return new IARSTM8();
     }
 
-    getGccFamilyCompilerPathForCpptools(): string | undefined {
+    getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined {
         //const gcc = File.fromArray([this.getToolchainDir().path, 'stm8', 'bin', `iccstm8${platform.exeSuffix()}`]);
         //return gcc.path;
         return undefined;
@@ -2487,8 +2490,10 @@ class MTI_GCC implements IToolchian {
         }
     }
 
-    getGccFamilyCompilerPathForCpptools(): string | undefined {
-        const gcc = File.fromArray([this.getToolchainDir().path, 'bin', this.getToolPrefix() + `gcc${platform.exeSuffix()}`]);
+    getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined {
+        const gcc = File.from(
+            this.getToolchainDir().path, 'bin', 
+            this.getToolPrefix() + `${type == 'c++' ? 'g++' : 'gcc'}${platform.exeSuffix()}`);
         return gcc.path;
     }
 
@@ -2760,8 +2765,10 @@ class RISCV_GCC implements IToolchian {
         }
     }
 
-    getGccFamilyCompilerPathForCpptools(): string | undefined {
-        const gcc = File.fromArray([this.getToolchainDir().path, 'bin', this.getToolPrefix() + `gcc${platform.exeSuffix()}`]);
+    getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined {
+        const gcc = File.from(
+            this.getToolchainDir().path, 'bin', 
+            this.getToolPrefix() + `${type == 'c++' ? 'g++' : 'gcc'}${platform.exeSuffix()}`);
         return gcc.path;
     }
 
@@ -2977,8 +2984,10 @@ class AnyGcc implements IToolchian {
         return new AnyGcc();
     }
 
-    getGccFamilyCompilerPathForCpptools(): string | undefined {
-        const gcc = File.fromArray([this.getToolchainDir().path, 'bin', this.getToolPrefix() + `gcc${platform.exeSuffix()}`]);
+    getGccFamilyCompilerPathForCpptools(type?: 'c' | 'c++'): string | undefined {
+        const gcc = File.from(
+            this.getToolchainDir().path, 'bin', 
+            this.getToolPrefix() + `${type == 'c++' ? 'g++' : 'gcc'}${platform.exeSuffix()}`);
         return gcc.path;
     }